requirement.txt
Django==1.11.8
celery==3.1.20
django-celery==3.1.17
django-environ==0.4.0
django-extensions==1.6.1
djangorestframework==3.3.0
django-cors-headers==1.1.0
paho-mqtt==1.3.1
pip==9.0.1
psycopg2==2.6.1
pycrypto==2.6.1
requests==2.18.4
wechatpy==1.6.0